requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pandas>=0.23.0
numpy<1.24
python-dotenv==0.19.0
webaiku @ git+https://github.com/dataiku/solutions-contrib.git@main#egg=webaiku&subdirectory=bs-infra
dataiku-api-client>=11.0.0

